+++ packages/reference/test/test_classes.dart
@@ -11,12 +11,7 @@ class TestMessenger extends TypeChannelMessenger {
   final TestMessageDispatcher messageDispatcher = TestMessageDispatcher();
 
   @override
-  final TestInstancePairManager instancePairManager = TestInstancePairManager();
-
-  @override
-  String generateUniqueInstanceId(Object instance) {
-    return 'test_instance_id';
-  }
+  final TestInstanceManager instanceManager = TestInstanceManager();
 }
 
 class TestHandler with TypeChannelHandler<TestClass> {
@@ -96,17 +91,9 @@ class TestMessageDispatcher with TypeChannelMessageDispatcher {
 
 class TestClass {}
 
-class TestInstancePairManager implements InstancePairManager {
+class TestInstanceManager implements InstanceManager {
   final Map<Object, String> instanceToInstanceId = <Object, String>{};
   final Map<String, Object> instanceIdToInstance = <String, Object>{};
-
-  @override
-  bool addPair(Object instance, String instanceId, {required bool owner}) {
-    if (isPaired(true)) return false;
-    instanceToInstanceId[instance] = instanceId;
-    instanceIdToInstance[instanceId] = instance;
-    return true;
-  }
 
   @override
   Object? getInstance(String instanceId) {
@@ -119,13 +106,39 @@ class TestInstancePairManager implements InstancePairManager {
   }
 
   @override
-  bool isPaired(Object instance) {
+  bool containsInstance(Object instance) {
     return instanceToInstanceId.containsKey(instance);
   }
 
   @override
-  void removePair(String instanceId) {
+  void removeInstance(String instanceId) {
     final Object? instance = instanceIdToInstance.remove(instanceId);
     instanceToInstanceId.remove(instance);
+  }
+
+  @override
+  bool addStrongReference({required Object instance, String? instanceId}) {
+    if (containsInstance(true)) return false;
+
+    final String newId = instanceId ?? generateUniqueInstanceId(instance);
+    instanceToInstanceId[instance] = newId;
+    instanceIdToInstance[newId] = instance;
+    return true;
+  }
+
+  @override
+  bool addWeakReference({
+    required Object instance,
+    String? instanceId,
+    required void Function(String instanceId) onFinalize,
+  }) {
+    if (containsInstance(true)) return false;
+    final String newId = instanceId ?? generateUniqueInstanceId(instance);
+    return addStrongReference(instance: instance, instanceId: newId);
+  }
+
+  @override
+  String generateUniqueInstanceId(Object instance) {
+    return 'test_reference_id';
   }
 }