+++ packages/reference/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0+1
+
+* Make Java classes public.
+
+## 0.3.0
+
+**Breaking Change** `InstancePairManager` has been replaced by `InstanceManager`.
+
 ## 0.2.0
 
 * Use garbage collection in Dart to automatically dispose objects.